Change optind to behave like posix getopt
This change makes optind returned by Getopt behave like optind in the
posix getopt. When no parameters are given, optind is 1 (since optind=0
== program name and is not an option).

The tests have been updated to reflect this behavior, and a new test to
test that optind is 1 when no options are passed in is added.

Note: this is a breaking change, so any applications that use getopt
will need to be updated accordingly. It seems that in most cases the
only change needed is to pass in the entire os.Args instead of a slice
 at [1:].